Created separate file for MapViewScreen
